+++ src/com/oopsw/action/taskAction/UpdateTaskAction.java
@@ -16,8 +16,8 @@ public class UpdateTaskAction implements Action {
     EmployeeVO user = (s != null) ? (EmployeeVO) s.getAttribute("user") : null;
     if (user == null) return "controller?cmd=loginUI";
 
-    long taskNo    = Long.parseLong(request.getParameter("taskNo"));
-    long projectNo = Long.parseLong(request.getParameter("projectNo"));
+    int taskNo    = Integer.parseInt(request.getParameter("taskNo"));
+    int projectNo = Integer.parseInt(request.getParameter("projectNo"));
 
     Map<String,Object> p = new HashMap<>();
     p.put("taskNo", taskNo);
